Move story translation crowdsourcing to its module
