+++ lib/modules/story_info/screens/story_info.dart
@@ -1,6 +1,6 @@
 import 'package:andax/modules/story_info/services/content_loader.dart';
 import 'package:andax/modules/story_editor/screens/narrative_editor.dart';
-import 'package:andax/screens/crowdsourcing_screen.dart';
+import 'package:andax/modules/story_translator/screens/crowdsourcing.dart';
 import 'package:andax/modules/store_play/screens/play.dart';
 import 'package:andax/widgets/loading_dialog.dart';
 import 'package:andax/widgets/rounded_back_button.dart';
